Added option --cgbs to let the user choose the cg_iteration_block_size

+++ Aliases.sh
@@ -251,16 +251,18 @@ if [ $LOAD_JOB_ALIASES = "TRUE" ]; then
 	        if [ "$CONFIRM" = "Y" ]; then break; elif [ "$CONFIRM" = "N" ]; then return; else  printf "\n\e[0;33m Please enter Y (yes) or N (no): \e[0m"; fi
         done
         
-        for BETA in b5.*; do
-	        echo $BETA
-	        cd $BETA
-	        for FILE in conf.????? conf.??????; do
-	            NUM=$(grep -o "[[:digit:]]*" <<< $FILE)
-                if [ ${NUM:+x} ]; then
-                    [ $(awk -v freq="$1" '{print $1%freq}' <<< $NUM) -ne 0 ] && rm -f $FILE ${FILE/conf/prng}
-                fi
-            done
-            cd ..
+        for BETA in b{5,6}.*; do
+			if [ -d $BETA ]; then
+				echo $BETA
+				cd $BETA
+				for FILE in conf.????? conf.??????; do
+					NUM=$(grep -o "[[:digit:]]*" <<< $FILE)
+					if [ ${NUM:+x} ]; then
+						[ $(awk -v freq="$1" '{print $1%freq}' <<< $NUM) -ne 0 ] && rm -f $FILE ${FILE/conf/prng}
+					fi
+				done
+				cd ..
+			fi
         done
     }
